adds the ability to save a sparse representation of the model
the saved VentralModel objects were huge because of the size of the
window tensors, plus the representations and all the other bits and
bobs. these aren't really necessary, I just need the initialized
version of the model when I load the metamer object back in, so I
added a work around. not sure if this is the most elegant, but it
works for now: we have a state_dict_sparse attribute that contains
just those attributes necessary to initialize the model (plus the
model name)

 - VentralModels now have state_dict_sparse and methods for
   saving/loading/intializing from it

 - Metamer.save now has a save_model_sparse boolean flag; if True, we
   try to save the state_dict_sparse attribute instead of the full
   model

 - Metamer.load now accepts a model_constructor argument, which we use
   if the model field of the saved Metamer object is a
   dictionary. added an example which should helpfully clarify things

 - adds tests for all the above

 - tests now make use of the tmp_path directive, which is super handy!