adds PoolingWindows class
this commit restructures the VentralStream models a bit: we split out
the bit that handles the creation of the pooling windows and their
actual pooling into a new class, PoolingWindows (in
canonical_computations/pooling.py). The idea is that this might be
helpful to other models and so we want to make it as useable as
possible

this required a bit of restructuring of code. that class also contains
the plotting functions, plot_windows and plot_window_sizes, and a new
forward method, but nothing else.

VentralStream models no longer have a windows attribute, though the
rest all still exist (and I may add windows back, if it's helpful)