handle weird-sized windows better
if the image we're giving to PoolingWindows is rectangular enough,
some of the windows will have nothing underneath them after being
cropped down to size. we can't necessarily remove them, so we just
make sure to handle it well. this means that any time we would divide
by that window (which would produce a np.nan since it's dividing by a
0), we replace that location in the output with 0.