adds rescale_factor
we now add rescale_factor, which scales up the text on the plots,
linewidth, and markersize, attempting to do so in a way that keeps
them easily visible. the resulting plots don't look great, but are at
least readable.

also corrects the creation of the figsize tuple for non-square
images (I was comparing the wrong parts of figsize and image.shape)

additional changes:

 - seed_image -> reference image