create_metamers.py can now be run from the command-line
with this, we run create_metamers.py from the command line. this
enables us to use bash's &> to redirect stdout/stderr to the log
file. contextlib.redirect_stdout/stderr seemed to be not update the
log file until the job finished running. for most rules, that's fine,
but since this rule in particular will run for a very long time, we'd
like to be able to check on it. now we can.

however, we can't make use of `shell:` because we'd like to be able to
run a bit of python code before and after the call to
cretae_metamers.py (in order to figure out which gpu to
use). onstart/onsuccess/onerror is close, but we can't pass a value
out of onstart, which is what I would need. so instead we use
os.system to just call create_metamers.py from the command line.

the downside of this is that snakemake won't have a helpful error
message if we hit an error, it will only know that the output wasn't
created. the upside is our log file will contain the error, which for
some reason it wasn't before