adds new attributes: saved_image/representation_gradient
with this, we add two new attributes, saved_image_gradient and
saved_representation_gradient, which save the gradients of the
synthesizing metamer-in-progress and the representation. these are
only stored if store_progress is not None (like saved_image and
saved_representation) and allow for easier debugging / investigating
of the model

this commit also overhauls how we save the model, now using a list of
keys to add to a growing save_dict. we also detach all tensors we
save, in order to avoid headaches coming from extra hooks or the like