handles zoom more cleverly
when creating 4k metamers, it doesn't make sense to keep scaling up
the videos and rep.png images, so I changed how I handle that. now,
there's a max reasonable zoom, 2 above the default (which is 256 for
RGC, 512 for V1). if the image is bigger than that, then we shrink it
down to a reasonable value. also cleans that all up a bit so that the
parts that are the same for V1 and RGC use the same blocks of
code. requires passing img_zoom through to a couple other functions