Make sure to pass the tensors we're storing to the cpu in synthesis
I was previously keeping the saved_image, saved_representation
tensors, as well as their gradients all on the device they were
computed on. this meant that, while running on 1 or more GPUs, we kept
them there, leading to increased memory demand on the GPU over time
for no gain, which made no sense. now we pass them all to the cpu as
we store them, for efficiency's sake

we also make sure that matched_rep, if it comes from
saved_representation, is on the same device as target_representation
during the representation_error method