overhauls how we handle changing image
I was having an issue where, with large images like we want to use for
the experiment, they were taking too long to update the content of the
ImageStim. in order to fix that, we now make use of PsychoPy's
StaticPeriod, which allows us to start a timer, execute some code, and
then finish off the chunk of time left. this allows us to pre-load the
next image during the off period of our experiment

as part of this, we clean up that janky next_stim_time calculation and
rename clock to expt_clock