overhauls V1 half-octave models
this commit changes how we create the PoolingWindows for the
half-octave bands. previously, PoolingWindows could take a list of
scales, including decimals, and would compute the sizes that
way (though this feature was undocumented). now, we revert to the
earlier way of doing things and require that num_scales be an
int. what we do instead is to create two separate PoolingWindows
objects, one based on the original image, one based on the
down-sampled one, and merge them.

 - adds PoolingWindows.merge method

 - change how we determine half_octave_img_res -- for *our* image
   sizes, this works better. in general, I think this is hard to
   figure out

 - PrimaryVisualCortex.scales attribute now goes coarse-to-fine (since
   this makes more sense)

 - Metamer coarse-to-fine synthesis expects the model's scales
   attribute to go coarse-to-fine rather than fine-to-coarse