moves where clamper gets initialized
now, it's after figuring out the device, to make sure that image is on
the proper device, for Two/FourMomentsClampers. This is only an issue
when using one gpu. If using none, then everything's on the cpu. If
using more than one, the reference image and metamer.matched_image
should both be on cpu (I think)