all targets supported
with this, all targets are now supported, which requires an overhaul
of how we handle getting and setting the relevant attributes.

Synthesis goes back to knowing nothing about the attribute structure
of its children classes, assuming it can just directly grab
self.matched_representation (instead of going through self.names). it
is the responsibility of the child class to make sure this is the
correct attribute to modify, which we do here by modifying the getter
and setter (and continued use of a self._names attribute)

 - adds way more documentation to mad competition

 - moves representation_error to Synthesis

 - mad competition has own representation_error method, to handle the
   multiple synthesis targets / models

 - normalizing the loss is now an option, not always done

 - many MADCompetition attributes now have "all" variants, which store
   across targets