adds choices for how to do coarse-to-fine optimization
standard way of doing coarse-to-fine (ctf) optimization is *not* what
I had been doing before. before this commit, ctf computed the gradient
with respect to the target scale individually, ignoring the
others (not doing anything to hold them fixed). the standard way is to
compute the gradient with respect to the target scale *and all coarser
scales*. this commit allows you to toggle between those two choices,
by setting coarse_to_fine to either 'together' or 'separate'.

also updates MADCompetition.save to appropriately save the related
attributes, and makes the tests for metamer and MAD more complete