removes foveal pixel representation
There were lingering issues with it: there was still some error in the
center of the image that optimization was not dealing with. I think the
way to handle this properly would be to fine tune the proportional size
of the representation (which I do in a hacky way by dividing by 10).
however, this foveal representation is not important for the scientific
questions we're interested in, so I'm removing it.

I am keeping some of the changes made as part of this though, namely,
making RGC's representation attribute a dictionary with a mean_luminance
key (and some of the resulting changes there). this makes interacting
with V1 and RGC models much more similar