removes cone from model-spec
with this commit, we no longer support the cone keyword in model
name (since we're removing that from plenoptic). this requires a fair
amount of change, but does make things simpler. still needs a bit of
testing

among other thigns: removes `image_name_for_expt`, which was just
image_name but removing 'cone_'.