removes postproc_metamers rule
all postproc_metamers was doing now was creating the gamma-corrected
metamer png file. unlike previously, when I had support for the cone
argument, there's no need to separate these two out. so we replaced that
rule with a gamma_correct_metamer rule that just drops the
gamma-corrected version in the right place.

related to this:

 - moved find_attempts into utils.py

 - generate_metamer_paths can now give either regular or
   gamma-corrected, and can find attempts as well