separates met vs met and met vs ref comparisons
with this, met vs met and met vs ref comparisons are not mixed in the
same experimental block. required a bit of restructuring of the code to
create presentation indices, as well as adding a new 'comp' key to the
path. finally, image set now includes 4 images instead of 2, so there
are now 2 of them instead of 4

will probably require some more changes moving forward