updates memory request on rusty
I was requesting the whole node when submitting GPU jobs, which is
inefficient. through ccn partition (no gpus), I get the whole node
regardless, and so I shouldn't request any memory (we'll have access to
all of it but slurm would kill the job if we went over). for jobs via
the gpu partition, however, I should specify the amount of memory