remove everything related to haploscope set up
we're running this on a single monitor, so this gets rid of everything
that referenced the haploscope / two screen / binocular set up.

biggest removal is of the ipd_calibration file, which is no longer
necessary

all the windows, text, and stimuli (everything drawn on the screen) were
lists, so that we had one for each screen. those are all single objects
now