adds computation and plot of orientation distribution
orientation anisotropy seems like a possible way of separating pink noise
from natural images (pink noise has equal power at all orientations). so
this adds a way to compute that and then plot it

we don't mean / median the vaules in each slice, because the
distributions are weird. instead we keep the full set of values and, for
now, we plot the median using catplot. might change